double rename task

+++ src/app/list/list.component.ts
@@ -24,8 +24,9 @@ export class ListComponent {
   constructor(private dataService: DataManagerService) {}
 
   delete(id: number) {
-    console.log(id);
-    this.dataService.deleteList(id);
+    if (confirm('Do you want to delete the list ' + this.list.name )) {
+      this.dataService.deleteList(id);
+    }  
   }
   newTask(ev) {
     const text = ev.target.value.trim();
@@ -36,47 +37,35 @@ export class ListComponent {
     }
   }
   addListSymbol(text) {
-    if (text !== '') {
-     
+    if (text !== '') {  
       this.dataService.addNewTask(text, this.list);
     }
     this.text = '';
-    
   }
   edit(node) {
    this.editingRe = true;
-    setTimeout(() => {
-     
+    setTimeout(() => { 
       node.focus();
     }, 0);
     this.editing = true;
-   
   }
   cancelEdit() {
     this.editing = false;
-    setTimeout(() => {
-     
+    setTimeout(() => {  
       this.addListSymbolRe(this.newName);
-    }, 0);
-    
+    }, 0); 
   }
   editName() {
   
     this.list.name = this.newName;
-   
     this.dataService.editListName(this.list);
     this.editing = false;
     this.editingRe = false;
   }
   addListSymbolRe(newName){
-  
-    if (newName!== ''){
-      
-      this.list.name = newName;
-      
-    }
-   
+    if (newName!== ''){  
+      this.list.name = newName;    
+    }  
     this.editingRe = false;
-
   }
 }
